Refactor: move WA parser datetime functions to its own file
